requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.5,<4.0
pymongo>=3.11,<4.0
pymongo[srv]>=3.11,<4.0
python-dotenv>=0.21.0
gunicorn>=20.1.0
certifi>=2020.12.5